remove assigns from resident Patient
