+++ HospitalManagementSystem/Views/AppointmentPatientDetailsView.xaml
@@ -119,8 +119,6 @@
                     <ListBox 
                     ItemsSource="{Binding DoctorsList}" 
                     Background="White"
-                    MouseDoubleClick="RemoveDoctorFromPatient"
-                    SelectedItem="{Binding ListSelectedDoctor}"
                     Foreground="{StaticResource PrimaryColor}" 
                     x:Name="DoctorsList"
                     DisplayMemberPath="Value"
@@ -128,56 +126,7 @@
                     <materialDesign:ColorZone Background="#0e263e" Grid.Row="1" Effect="{DynamicResource MaterialDesignShadowDepth5}">
                     <TextBlock Text="{Binding DoctorsNumber}" Foreground="White" Margin="16"/>
                 </materialDesign:ColorZone>
-                <Button Style="{StaticResource MaterialDesignFloatingActionMiniAccentButton}"                                
-                        Command="{x:Static materialDesign:DialogHost.OpenDialogCommand}"
-                        VerticalAlignment="Bottom" HorizontalAlignment="Right" 
-                        Grid.Row="0" Margin="0 0 28 -20"
-                        BorderThickness="0"
-                        BorderBrush="Transparent"
-                        materialDesign:ShadowAssist.ShadowDepth="Depth0"
-                        Background="{StaticResource PrimaryColor}">
-                            <Button.CommandParameter>
-                                <StackPanel Margin="16">
-                                    <ComboBox 
-                                        DisplayMemberPath="Value"
-                                        SelectedValuePath="Key"
-                                        SelectedItem="{Binding DoctorComboBox}"
-                                        ItemsSource="{Binding DoctorsComboBox}" 
-                                        IsEditable="True"
-                                        materialDesign:HintAssist.Hint="Add Doctor"
-                                        Style="{StaticResource MaterialDesignFloatingHintComboBox}"  
-                                        Margin="0 8 0 0" 
-                                        HorizontalAlignment="Left"
-                                        Width="200"
-                                        x:Name="DoctorsComboBox"/>
-                                    <StackPanel Orientation="Horizontal" HorizontalAlignment="Right" >
-                                            <Button Style="{StaticResource MaterialDesignFlatButton}"
-                                                IsDefault="True"
-                                                Margin="0 8 8 0"
-                                                Command="{Binding assignDoctor }">
-                                                <Button.CommandParameter>
-                                                <system:Boolean>True</system:Boolean>
-                                            </Button.CommandParameter>
-                                            SUBMIT
-                                        </Button>
-                                        <Button Style="{StaticResource MaterialDesignFlatButton}"
-                                                IsCancel="True"
-                                                Margin="0 8 8 0"
-                                                Command="materialDesign:DialogHost.CloseDialogCommand">
-                                            <Button.CommandParameter>
-                                                <system:Boolean>False</system:Boolean>
-                                            </Button.CommandParameter>
-                                            CANCEL
-                                        </Button>
-                                    </StackPanel>
-                                </StackPanel>
-                            </Button.CommandParameter>
-                    <Viewbox Width="22" Height="22">
-                        <Canvas Width="24" Height="24">
-                            <Path Data="M19,13H13V19H11V13H5V11H11V5H13V11H19V13Z" Fill="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType=Button}, Path=Foreground}" />
-                        </Canvas>
-                    </Viewbox>
-                </Button>
+                
             </Grid>
         </Border>
     </materialDesign:DialogHost>
